+++ src/main/java/icons/KdbIcons.java
@@ -59,6 +59,7 @@ public final class KdbIcons {
         public static final @NotNull Icon ToolCrosshair = load("/org/kdb/inside/brains/icons/chart/tool_crosshair.svg");
         public static final @NotNull Icon ToolPoints = load("/org/kdb/inside/brains/icons/chart/tool_points.svg");
         public static final @NotNull Icon ToolMeasure = load("/org/kdb/inside/brains/icons/chart/tool_measure.svg");
+        public static final @NotNull Icon ToolLegend = load("/org/kdb/inside/brains/icons/chart/tool_legend.svg");
 
         public static final @NotNull Icon Templates = load("/org/kdb/inside/brains/icons/chart/template.svg");
     }