+++ src/main/java/org/kdb/inside/brains/view/chart/tools/ToolMode.java
@@ -1,0 +1,15 @@
+package org.kdb.inside.brains.view.chart.tools;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public interface ToolMode {
+    @NotNull
+    String name();
+
+    @NotNull
+    String getText();
+
+    @Nullable
+    String getDescription();
+}